+++ src/agent/opts.js
@@ -6,6 +6,7 @@ exports.host = '127.0.0.1';
 exports.port = 42699;
 exports.serverHeader = 'Instana Agent';
 exports.agentUuid = undefined;
+exports.extraHttpHeadersToCapture = [];
 
 
 exports.init = function init(config) {