+++ test/tracing/httpServer/app.js
@@ -15,7 +15,7 @@ var express = require('express');
 var morgan = require('morgan');
 
 var app = express();
-var logPrefix = 'Express HTTP client: Server (' + process.pid + '):\t';
+var logPrefix = 'Express HTTP: Server (' + process.pid + '):\t';
 
 if (process.env.WITH_STDOUT) {
   app.use(morgan(logPrefix + ':method :url :status'));
@@ -24,14 +24,7 @@ if (process.env.WITH_STDOUT) {
 app.use(bodyParser.json());
 
 app.get('/', function(req, res) {
-  res.set('Foobar', '42');
   res.sendStatus(200);
-});
-
-app.get('/timeout', function(req, res) {
-  setTimeout(function() {
-    res.sendStatus(200);
-  }, 1000);
 });
 
 app.listen(process.env.APP_PORT, function() {