+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
  - Include additional debugging data in log when data transmission to agent fails.
+ - Support recording of user-configurable HTTP headers in HTTP client and server instrumentation.
 
 ## 1.39.0
  - Include reporting PID in agent logs.